Join Request functionality working for strict rooms
